Move string escaping to a new Strings module
