+++ jproto/src/main/java/uk/bs338/hashLisp/jproto/hons/PrettyPrinter.java
@@ -17,30 +17,6 @@ public class PrettyPrinter<V extends IValue> {
         this.heap = heap;
         stringTag = heap.makeSymbol("*string");
     }
-
-    private @NotNull IntStream quoteChar(int ch) {
-        /* Java backslash sequences are \t, \b, \n, \r, \f, \', \", \\ */
-        return switch (ch) {
-            case '\t' -> IntStream.of('\\', 't');
-            case '\b' -> IntStream.of('\\', 'b');
-            case '\n' -> IntStream.of('\\', 'n');
-            case '\r' -> IntStream.of('\\', 'r');
-            case '\f' -> IntStream.of('\\', 'f');
-            case '\'' -> IntStream.of('\\', '\'');
-            case '"' -> IntStream.of('\\', '"');
-            case '\\' -> IntStream.of('\\', '\\');
-            default ->
-                /* NB. doesn't quote non-printable */
-                IntStream.of(ch);
-        };
-    }
-    
-    public @NotNull String quoteString(@NotNull String str) {
-        /* Slow(?) but elegant */
-        var escaped = str.codePoints().flatMap(this::quoteChar);
-        int[] result = IntStream.concat(IntStream.of('"'), IntStream.concat(escaped, IntStream.of('"'))).toArray();
-        return new String(result, 0, result.length);
-    }
     
     private @NotNull Optional<String> stringifyNonList(@NotNull V val) {
         if (!val.isConsRef())
@@ -51,7 +27,7 @@ public class PrettyPrinter<V extends IValue> {
             if (uncons.fst().isSymbolTag())
                 return Optional.of(listAsString(heap, uncons.snd()));
             if (uncons.fst().equals(stringTag))
-                return Optional.of(quoteString(listAsString(heap, uncons.snd())));
+                return Optional.of(Strings.quoteString(listAsString(heap, uncons.snd())));
         }
         catch (IllegalStateException e) {
             return Optional.of(val.toString());
