+++ jproto/src/main/java/uk/bs338/hashLisp/jproto/reader/CharClassifier.java
@@ -67,21 +67,4 @@ public class CharClassifier {
     public EnumSet<CharClass> classifyChar(@NotNull String ch) {
         return classifyChar(ch.codePointAt(0));
     }
-    
-    public @NotNull String interpretEscapedChar(@NotNull String ch) {
-        /* Java backslash sequences are \t, \b, \n, \r, \f, \', \", \\ */
-        return switch (ch) {
-            case "t" -> "\t";
-            case "b" -> "\b";
-            case "n" -> "\n";
-            case "r" -> "\r";
-            case "f" -> "\f";
-            case "'" -> "'";
-            case "\"" -> "\"";
-            case "\\" -> "\\";
-            default ->
-                /* We're liberal here and return accept anything */
-                ch;
-        };
-    }
 }
